refactor: move endpoint logic into handlers
