+++ server/state/test-state.json
@@ -1,33 +1,1 @@
-{
-    "last_updated": "2024-04-24T22:17:59+01:00",
-    "portfolios": [
-        {
-            "name": "test",
-            "uuid": "test-portfolio-id",
-            "type": "test",
-            "deleted": false,
-            "current_strategy": {
-                "name": "HODL",
-                "currency": "ETH",
-                "open_offers": [
-                    {
-                        "client_order_id": "test",
-                        "product_id": "GBP-ETH",
-                        "side": "BUY",
-                        "config": {
-                            "type": "limit_limit_gtd",
-                            "base_size": "10",
-                            "limit_price": "10",
-                            "post_only": true,
-                            "end_time": "2024-04-24T22:22:58+01:00"
-                        },
-                        "self_trade_prevention_id": "test",
-                        "retail_portfolio_id": "test-portfolio-id"
-                    }
-                ],
-                "closed_offers": null
-            },
-            "previous_strategies": null
-        }
-    ]
-}
+{"last_updated":"2024-04-27T17:36:48+01:00","portfolios":[{"name":"test","uuid":"test-portfolio-id","type":"test","deleted":false,"current_strategy":{"name":"HODL","currency":"ETH","open_offers":[{"client_order_id":"test","product_id":"GBP-ETH","side":"BUY","config":{"type":"limit_limit_gtd","base_size":"10","limit_price":"10","post_only":true,"end_time":"2024-04-27T17:41:47+01:00"},"self_trade_prevention_id":"test","retail_portfolio_id":"test-portfolio-id"}],"closed_offers":null},"previous_strategies":null}]}
